split files and try with a blow demo
