+++ tsconfig.json
@@ -14,6 +14,7 @@
     "lib": ["es2019", "dom"],
     "types": ["@webgpu/types"],
     "baseUrl": "./src/",
+    "allowImportingTsExtensions": true,
     "verbatimModuleSyntax": false
   },
   "include": ["./src/*"]
